New DB columns, populating import

+++ app/src/main/java/com/tfcode/comparetout/model/importers/alphaess/AlphaESSTransformedData.java
@@ -17,7 +17,6 @@
 package com.tfcode.comparetout.model.importers.alphaess;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 
@@ -36,13 +35,13 @@ public class AlphaESSTransformedData {
     @ColumnInfo(defaultValue = "0")
     private double charge = 0;
     @ColumnInfo(defaultValue = "NULL")
-    private Integer millisSinceEpoch;
+    private Long millisSinceEpoch;
 
-    public Integer getMillisSinceEpoch() {
+    public Long getMillisSinceEpoch() {
         return millisSinceEpoch;
     }
 
-    public void setMillisSinceEpoch(int millisSinceEpoch) {
+    public void setMillisSinceEpoch(Long millisSinceEpoch) {
         this.millisSinceEpoch = millisSinceEpoch;
     }
